requirements.txt
discord.py
requests==2.27.1
Pillow==9.1.1